+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 
 -   #32 : Add support for `nvcc` Compiler and `cuda` language as a possible option.
 -   #48 : Fix incorrect handling of imports in `cuda`.
+-   #42 : Add support for custom kernel in`cuda`.
+-   #42 : Add Cuda module to Pyccel. Add support for `cuda.synchronize` function.
 
 ## \[UNRELEASED\]